backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1

# InfluxDB
influxdb==5.3.2

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-multipart==0.0.6

# Encryption
cryptography==42.0.1

# Garmin Connect (compatible versions)
garminconnect==0.2.24    
garth==0.4.47

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
